requirements.txt
openai-whisper
transformers
sentencepiece
accelerate
yt-dlp
webvtt-py
